A patch of size 44Kbytes... in short:
Found and repaired the off-by-one error in rrd_fetch_fn().
As a result I had to remove the hacks in rrd_fetch_fn(),
rrd_tool.c, vdef_calc(), data_calc(), data_proc() and
reduce_data().  There may be other places which I didn't
find so be careful.

Enhanced debugging in rrd_fetch_fn(), it shows the RRA selection
process.

Added the ability to print VDEF timestamps.  At the moment it
is a hack, I needed it now to fix the off-by-one error.
If the format string is "%c" (and nothing else!), the time
will be printed by both ctime() and as a long int.

Moved some code around (slightly altering it) from rrd_graph()
  initializing     now in rrd_graph_init()
  options parsing  now in rrd_graph_options()
  script parsing   now in rrd_graph_script()


git-svn-id: svn://svn.oetiker.ch/rrdtool/trunk/program@59 a5681a0c-68f1-0310-ab6d-d61299d08faa