Bug #429: Clean-up of reserved identifier violations
Some identifiers did not fit to the expected naming convention of the
C language standard. This detail was fixed by the deletion of leading underscores.
The probability for name clashes was reduced for the affected include guards
by the addition of a kind of universally unique identifier as a suffix.

Signed-off-by: Markus Elfring <elfring@users.sourceforge.net>