autogen.sh: use POSIX comparison for test
Although bash recognizes ==, and the shell has a shebang to use it, some
of us still automatically type "sh autogen.sh" instead of
"./autogen.sh", and this breaks on debian/jessie where /bin/sh is not
bash.  The fix is trivial and should be safe for bash too (which
recommends using = instead of == with the test command for POSIX
conformance.