automake: update dist checks for recent changes
Various files have been added and removed causing the dist and distcheck
Make targets to fail.  This commit fixes dist; distcheck is improved but
is not complete because the bindings Makefile.am assumes the build is
performed in the source tree.  (Standard autotool practice now is to
require support for separated builds.)

Signed-off-by: Peter A. Bigot <pab@pabigot.com>