autoconf: rework config file locations
If we assume somebody puts the src directory in the include file path to
CC, we don't need relative symlinks to rrd_config.h which will make
separated builds much easier to support.  Then we don't need to have the
config files in the root directory and symlinks back and forth between
there and src.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>