rrd_daemon: move include of arpa/inet.h to avoid use on Windows
arpa/inet.h may not be available on Windows systems; since I can't test
that, avoid trying to include it there.  Since the code that depends on
it is conditional on a define that would be provided by it, there should
be no impact from its absence except that somebody who uses Windows
would have to rework the fix for ipv4-over-ipv6 for that platform.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>