misc doc: fix erronous use of section qualifier in intra-package links
rrdxport and rrdcached do belong in section 1 of the Unix manpages, but
they are documented within the RRDTool hierarchy and use of the
qualifier prevents them from being found.  Note also the need to
override the substitute text or modify the context so that the generated
HTML content reads correctly.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>