Allow empty imginfo format string
Empty -f|--imginfo used to be very usefull to print data as text format
on stdout. It was allowed up to 1.4.7 and was particularly interesting
because there was no filename/width/height printed out. The "feature"
was broken in 1.4.8 by the addition of a format check on the imginfo
value. The recent rework based on regexp still breaks it. Adding it back
would preserve compatibility on scripts/parser eating rrdtool graph
/dev/null -f '' output.

Before :

[root@nagios-2b ~]# rrdtool graph /dev/null -f "" --start -1d \
    DEF:d=/var/lib/rrd/nagios-2b/load.rrd:load5:AVERAGE \
    VDEF:v=d,AVERAGE \
    PRINT:v:%.6lf
5.284605

After :

[root@nagios-2b ~]# rrdtool graph /dev/null -f "" --start -1d \
    DEF:d=/var/lib/rrd/nagios-2b/load.rrd:load5:AVERAGE \
    VDEF:v=d,AVERAGE \
    PRINT:v:%.6lf
ERROR: bad format for imginfo