RRD client: Clean up the rrd_client*_is_connected() functions.
Rename rrd_client_is_any_connected() to rrd_client_is_connected() and replace
the old rrd_client_is_connected() with rrd_client_address(). That is, instead
of specifying the address to check for (which only really makes sense for the
global default which is used in arbitrary places), add a way to retrieve the
server address.