rrd_open() should position "to the first cdp in the first rra"
but RRD_READVALUES was buggy.  It left the fd positioned at the end of the
file, but with rrd_file->pos still set to just after the header.
(rados, which doesn't use the fd, would be unafffected by this bug, leading
to differing behaviour).

The comment for rrd_open implies we should rrd_seek() back to just after
the header.  So let's resolve the inconsistency that way.

The old code with d_offset was completely pointless.  It was used to
restore rrd_file->pos, but rrd_file->pos was not changed, even by the
__rrd_read macro magic.