fadvise and madvise code in rrd_open can be kept together
There was no reason to break up fadvise and madvise.

Putting them together makes it _slightly_ clearer that the exising #ifdefs
are intended.  (Documentation on the effect of fadvise() on mmap() is
hard to find, so it was simpler to avoid doing that).

Also, setvbuf() clearly doesn't apply to fd's.  Buffering is a function of
stdio e.g. fread(), fwrite() etc., but rrd_open() does not (currently?)
use these, it uses fds directly.  Hence, we have no buffering already :).

There must be some reason people use fread() but disable buffering,
but I can't think why rrd_open() might want to do that in future.
Remove the commented code which uses setvbuf().