rrd_open() does not want to enable drop-behind of memory-mapped header
MADV_SEQUENTIAL mentions drop-behind.  Override it for the header now
we've read it, in case anyone implemented drop-behind.

Linux does not currently implement drop-behind.  My performance test
confirmed this did not help, but it didn't hurt either.

Do *not* fall back to fadvise() for !HAVE_MMAP.  We've already copied
the header in that case.  Doing e.g. FADV_NORMAL on Linux (4.12) on
*any* region would negate the effect of previous FADV_SEQUENTIAL.