fix incorrect use of snprintf in rrd_cgi
DS_NAM_SIZE is used to size a buffer, but there is no relation to ds_nam.
Fortunately this lead to the buffer being over-sized.

Replace with code which does not rely on calculating the size of a
formatted string.  (If the error message is around the maximum possible
size supported by librrd, we might drop a closing ']'.  That won't cause
us a problem).

We don't have a shim for asprintf().  rrd_asprintf() and friends are for
writing e.g. XML etc without locale-specific formatting.