Update MSVC .vcxproj files (SAFESH, IntDir)
Win32 Debug configurations have EditAndContinue set, whereas /SAFESEH
is incompatible with EditAndContinue. Add
<ImageHasSafeExceptionHandlers>false</ImageHasSafeExceptionHandlers>
to these configurations.

- Fixes the following warning in Visual Studio build log files, e.g.:
  warning LNK4075: ignoring '/EDITANDCONTINUE' due to
  '/SAFESEH' specification

The projects librrd-8, rrdcgi, rrdtool and rrdupdate are built in
the same directory, which is OK, as there is no overlap. No need for a
separate IntDir for each project. To silence the warning, add:
<IntDirSharingDetected>None</IntDirSharingDetected>

- Fixes the following warning in Visual Studio build log files, e.g.:
  warning MSB8028: The intermediate directory (Debug\) contains files
  shared from another project (librrd-8.vcxproj).
  This can lead to incorrect clean and rebuild behavior.