rrdcached dump support (#1235)
* add missing newlines (\n) for TUNE

Help strings for TUNE command were lacking the trailing newline
character. This causes the FLUSH syntax printed in the same line as
TUNE's.

* fix sendall edge cases

The `sendall` func is implemented as a loop over multiple send() calls until all bytes are sent. Previously on subsequent calls the incorrect buffer was used for sending.
This change also forbids `allow_retry` when at least one byte was successfully sent already. This would otherwise mess up the protocol, when retrying from the start.

* rrcached: implement DUMP

* update tests for remote dump support

Previously, tests replaced the daemon-based dump use with a local use.
This is changed in this commit as dump is now supported via daemon.

* update doc/rrddump

* fix format argument type

* update to latest windows vcpkg tag